# Conflicts:
#	nts.uk/uk.com/com.web/nts.uk.com.web/src/main/webapp/view/ccg/share/ccg.vm.ts
#	nts.uk/uk.com/query/nts.uk.query.app/src/main/java/nts/uk/query/app/employee/RegulationInfoEmpQueryDto.java
#	nts.uk/uk.com/query/nts.uk.query.app/src/main/java/nts/uk/query/app/employee/RegulationInfoEmployeeFinder.java
#	nts.uk/uk.com/query/nts.uk.query.infra/src/main/java/nts/uk/query/infra/repository/employee/JpaRegulationInfoEmployeeRepository.java
#	nts.uk/uk.com/query/nts.uk.query.model/src/main/java/nts/uk/query/model/employee/RegulationInfoEmployeeRepository.java

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.app/src/main/java/nts/uk/ctx/at/record/app/command/remainingnumber/empinfo/basicinfo/update/UpdateSpLea6InfoCommandHandler.java
@@ -14,11 +14,11 @@ import nts.uk.shr.pereg.app.command.PeregUpdateCommandHandler;
 
 @Stateless
 public class UpdateSpLea6InfoCommandHandler extends CommandHandler<UpdateSpecialleave6informationCommand>
-implements PeregUpdateCommandHandler<UpdateSpecialleave6informationCommand>{
+		implements PeregUpdateCommandHandler<UpdateSpecialleave6informationCommand> {
 
-	@Inject 
+	@Inject
 	private SpLeaInfoCommandHandler updateSpLeaInfoCommandHandler;
-	
+
 	@Override
 	public String targetCategoryCd() {
 		return "CS00030";
@@ -31,11 +31,12 @@ implements PeregUpdateCommandHandler<UpdateSpecialleave6informationCommand>{
 
 	@Override
 	protected void handle(CommandHandlerContext<UpdateSpecialleave6informationCommand> context) {
-		
+
 		val command = context.getCommand();
 		String cid = AppContexts.user().companyId();
-		SpecialLeaveBasicInfo domain = new SpecialLeaveBasicInfo(cid,command.getSID(),SpecialLeaveCode.CS00030.value, command.getUseAtr().intValue(), command.getAppSet().intValue(), command.getGrantDate(),
-				command.getGrantDays()!=null? command.getGrantDays().intValue() : null, command.getGrantTable());
+		SpecialLeaveBasicInfo domain = new SpecialLeaveBasicInfo(cid, command.getSID(), SpecialLeaveCode.CS00030.value,
+				command.getUseAtr(), command.getAppSet(), command.getGrantDate(),
+				command.getGrantDays() != null ? command.getGrantDays().intValue() : null, command.getGrantTable());
 		updateSpLeaInfoCommandHandler.updateHandler(domain);
 	}
